refactor: #9 Local Layer mapper 분리
