+++ local/src/main/java/com/kdjj/local/dataSource/RecipeListLocalDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.kdjj.local.dataSource
 import com.kdjj.data.datasource.RecipeListLocalDataSource
 import com.kdjj.domain.model.Recipe
 import com.kdjj.local.dao.RecipeListDao
-import com.kdjj.local.dto.toDomain
+import com.kdjj.local.mapper.toDomain
 import javax.inject.Inject
 
 internal class RecipeListLocalDataSourceImpl @Inject constructor(
