+++ local/src/main/java/com/kdjj/local/dataSource/RecipeTypeLocalDataSourceImpl.kt
@@ -3,8 +3,8 @@ package com.kdjj.local.dataSource
 import com.kdjj.data.datasource.RecipeTypeLocalDataSource
 import com.kdjj.domain.model.RecipeType
 import com.kdjj.local.dao.RecipeTypeDao
-import com.kdjj.local.dto.toDomain
-import com.kdjj.local.dto.toDto
+import com.kdjj.local.mapper.toDomain
+import com.kdjj.local.mapper.toDto
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
