+++ local/src/main/java/com/kdjj/local/dto/RecipeDto.kt
@@ -2,7 +2,6 @@ package com.kdjj.local.dto
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.kdjj.domain.model.Recipe
 
 internal data class RecipeDto(
     @Embedded val recipeMeta: RecipeMetaDto,
@@ -17,19 +16,3 @@ internal data class RecipeDto(
     )
     val steps: List<RecipeStepDto>
 )
-
-internal fun RecipeDto.toDomain(): Recipe =
-    Recipe(
-        recipeId = recipeMeta.recipeMetaId,
-        title = recipeMeta.title,
-        type = recipeType.toDomain(),
-        stuff = recipeMeta.stuff,
-        imgPath = recipeMeta.imgPath,
-        stepList = steps.map { it.toDomain() },
-        authorId = recipeMeta.authorId,
-        viewCount = 0,
-        isFavorite = recipeMeta.isFavorite,
-        createTime = recipeMeta.createTime,
-        state = recipeMeta.state
-    )
-
