+++ local/src/main/java/com/kdjj/local/dto/RecipeTempMetaDto.kt
@@ -4,7 +4,6 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import com.kdjj.domain.model.Recipe
 import com.kdjj.domain.model.RecipeState
 
 @Entity(
@@ -31,16 +30,3 @@ internal data class RecipeTempMetaDto(
     @ColumnInfo(defaultValue = "1")
     val recipeTypeId: Long
 )
-
-internal fun Recipe.toTempDto(): RecipeTempMetaDto =
-    RecipeTempMetaDto(
-        recipeId,
-        title,
-        stuff,
-        imgPath,
-        authorId,
-        isFavorite,
-        createTime,
-        state,
-        type.id.toLong()
-    )
