+++ local/src/main/java/com/kdjj/local/mapper/RecipeTempMetaDtoMapper.kt
@@ -1,0 +1,17 @@
+package com.kdjj.local.mapper
+
+import com.kdjj.domain.model.Recipe
+import com.kdjj.local.dto.RecipeTempMetaDto
+
+internal fun Recipe.toTempDto(): RecipeTempMetaDto =
+    RecipeTempMetaDto(
+        recipeId,
+        title,
+        stuff,
+        imgPath,
+        authorId,
+        isFavorite,
+        createTime,
+        state,
+        type.id.toLong()
+    )
