+++ local/src/main/java/com/kdjj/local/mapper/RecipeTypeDtoMapper.kt
@@ -1,0 +1,16 @@
+package com.kdjj.local.mapper
+
+import com.kdjj.domain.model.RecipeType
+import com.kdjj.local.dto.RecipeTypeDto
+
+internal fun RecipeTypeDto.toDomain(): RecipeType =
+    RecipeType(
+        recipeTypeId.toInt(),
+        title
+    )
+
+internal fun RecipeType.toDto(): RecipeTypeDto =
+    RecipeTypeDto(
+        id.toLong(),
+        title
+    )
